pyproject.toml
[tool.poetry]
name = "papyrusbot"
version = "0.1.0"
description = "manage your papyrus scroll"
authors = ["OpenAlcoholics <team@openalcoholics.group>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0"
requests = "^2.28.1"
pydantic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
